+++ platform_agnostic/src/main/java/com/github/mdcdi1315/mdex/mixin/ServerLevelAccessor.java
@@ -1,0 +1,21 @@
+package com.github.mdcdi1315.mdex.mixin;
+
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.level.CustomSpawner;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Mutable;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+import java.util.List;
+
+@Mixin(ServerLevel.class)
+public interface ServerLevelAccessor
+{
+    @Accessor("customSpawners")
+    List<CustomSpawner> GetCustomSpawners();
+
+    @Mutable
+    @Accessor("customSpawners")
+    void SetCustomSpawners(List<CustomSpawner> s);
+}