+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-### Version 2.2.2:
+### Version 2.2.3:
 
--> Minor bugfixes, port to BML 1.0.15
+-> Added a new way to spawn mobs in Mining Dimension: The Aggressive Spawner. It helps to spawn mobs uncapped.
+
+-> Added a new furnace variant based on the Hardstone block
+
+-> Reimplemented the Loot subsystem
+
+-> Stone tools can be now created by using the Hardstone block
+
+-> Minor bugfixes